rename logger to log
