+++ packages/cli/src/commands/init.ts
@@ -3,7 +3,7 @@ import { Command, Flags } from '@oclif/core'
 import * as fs from 'fs'
 import * as path from 'path'
 
-import log from '../logger'
+import log from '../log'
 
 export default class Init extends Command {
   static override description = 'Initializes a new Mimic-compatible project structure in the specified directory'
